refactor(page-builder): modularize state actions
