+++ packages/ui/src/components/cms/page-builder/state/actions/add.js
@@ -1,0 +1,9 @@
+import { addComponent } from "./helpers";
+import { commit } from "../history";
+export function add(state, action) {
+  return commit(
+    state,
+    addComponent(state.present, action.parentId, action.index, action.component),
+  );
+}
+export default add;
