+++ packages/ui/src/components/cms/page-builder/state/actions/duplicate.js
@@ -1,0 +1,6 @@
+import { duplicateComponent } from "./helpers";
+import { commit } from "../history";
+export function duplicate(state, action) {
+  return commit(state, duplicateComponent(state.present, action.id));
+}
+export default duplicate;
