+++ packages/ui/src/components/cms/page-builder/state/actions/duplicate.ts
@@ -1,0 +1,10 @@
+import type { HistoryState } from "@acme/types";
+import type { DuplicateAction } from "./types";
+import { duplicateComponent } from "./helpers";
+import { commit } from "../history";
+
+export function duplicate(state: HistoryState, action: DuplicateAction): HistoryState {
+  return commit(state, duplicateComponent(state.present, action.id));
+}
+
+export default duplicate;
