+++ packages/ui/src/components/cms/page-builder/state/actions/index.js
@@ -1,0 +1,8 @@
+export { add } from "./add";
+export { move } from "./move";
+export { remove } from "./remove";
+export { duplicate } from "./duplicate";
+export { update } from "./update";
+export { resize } from "./resize";
+export { set } from "./set";
+export { setGridCols } from "./set-grid-cols";
