+++ packages/ui/src/components/cms/page-builder/state/actions/index.ts
@@ -1,0 +1,23 @@
+export type {
+  AddAction,
+  MoveAction,
+  RemoveAction,
+  DuplicateAction,
+  UpdateAction,
+  ResizeAction,
+  SetAction,
+  ChangeAction,
+  UndoAction,
+  RedoAction,
+  SetGridColsAction,
+  Action,
+} from "./types";
+
+export { add } from "./add";
+export { move } from "./move";
+export { remove } from "./remove";
+export { duplicate } from "./duplicate";
+export { update } from "./update";
+export { resize } from "./resize";
+export { set } from "./set";
+export { setGridCols } from "./set-grid-cols";
