+++ packages/ui/src/components/cms/page-builder/state/actions/move.js
@@ -1,0 +1,6 @@
+import { moveComponent } from "./helpers";
+import { commit } from "../history";
+export function move(state, action) {
+  return commit(state, moveComponent(state.present, action.from, action.to));
+}
+export default move;
