+++ packages/ui/src/components/cms/page-builder/state/actions/move.ts
@@ -1,0 +1,10 @@
+import type { HistoryState } from "@acme/types";
+import type { MoveAction } from "./types";
+import { moveComponent } from "./helpers";
+import { commit } from "../history";
+
+export function move(state: HistoryState, action: MoveAction): HistoryState {
+  return commit(state, moveComponent(state.present, action.from, action.to));
+}
+
+export default move;
