+++ packages/ui/src/components/cms/page-builder/state/actions/remove.js
@@ -1,0 +1,6 @@
+import { removeComponent } from "./helpers";
+import { commit } from "../history";
+export function remove(state, action) {
+  return commit(state, removeComponent(state.present, action.id));
+}
+export default remove;
