+++ packages/ui/src/components/cms/page-builder/state/actions/set-grid-cols.js
@@ -1,0 +1,4 @@
+export function setGridCols(state, action) {
+  return { ...state, gridCols: action.gridCols };
+}
+export default setGridCols;
