+++ packages/ui/src/components/cms/page-builder/state/actions/set-grid-cols.ts
@@ -1,0 +1,11 @@
+import type { HistoryState } from "@acme/types";
+import type { SetGridColsAction } from "./types";
+
+export function setGridCols(
+  state: HistoryState,
+  action: SetGridColsAction,
+): HistoryState {
+  return { ...state, gridCols: action.gridCols };
+}
+
+export default setGridCols;
