+++ packages/ui/src/components/cms/page-builder/state/actions/set.js
@@ -1,0 +1,5 @@
+import { commit } from "../history";
+export function set(state, action) {
+  return commit(state, action.components);
+}
+export default set;
