+++ packages/ui/src/components/cms/page-builder/state/actions/update.js
@@ -1,0 +1,6 @@
+import { updateComponent } from "./helpers";
+import { commit } from "../history";
+export function update(state, action) {
+  return commit(state, updateComponent(state.present, action.id, action.patch));
+}
+export default update;
