+++ packages/ui/src/components/cms/page-builder/state/index.d.ts
@@ -1,7 +1,5 @@
-import type { HistoryState } from "@acme/types";
-import type { Action } from "./actions";
 export type { Action } from "./actions";
 export { historyStateSchema } from "./schema";
-export declare function reducer(state: HistoryState, action: Action): HistoryState;
-export default reducer;
+export { reducer } from "./reducer";
+export { reducer as default } from "./reducer";
 //# sourceMappingURL=index.d.ts.map
