+++ packages/ui/src/components/cms/page-builder/state/reducer.d.ts
@@ -1,0 +1,5 @@
+import type { HistoryState } from "@acme/types";
+import type { Action } from "./actions";
+export declare function reducer(state: HistoryState, action: Action): HistoryState;
+export default reducer;
+//# sourceMappingURL=reducer.d.ts.map
